backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.8.2
chromadb==0.5.5
pypdf==5.1.0
nltk==3.9.1
python-multipart
